+++ src/platform_impl/macos/monitor.rs
@@ -235,7 +235,7 @@ impl MonitorHandle {
       Some(screen) => screen,
       None => return 1.0, // default to 1.0 when we can't find the screen
     };
-    NSScreen::backingScaleFactor(&screen) as f64
+    NSScreen::backingScaleFactor(&screen)
   }
 
   pub fn video_modes(&self) -> impl Iterator<Item = RootVideoMode> {
